Added linters, editorconfig and truffle ^5.0
