+++ .editorconfig
@@ -1,0 +1,17 @@
+# top-most EditorConfig file
+root = true
+
+# Unix-style newlines with a newline ending every file
+[*]
+end_of_line = lf
+insert_final_newline = true
+
+# Styles
+[*]
+indent_style = space
+
+[*.js]
+indent_size = 4
+
+[*.sol]
+indent_size = 2
