+++ .eslintignore
@@ -1,0 +1,3 @@
+aztec-crypto-js/
+docs/
+node_modules/
