+++ .soliumignore
@@ -1,0 +1,2 @@
+aztec-crypto-js
+node_modules
