+++ .soliumrc.json
@@ -1,0 +1,21 @@
+{
+  "extends": "solium:all",
+  "plugins": ["security"],
+  "rules": {
+    "blank-lines": "off",
+    "error-reason": "off",
+    "indentation": ["error", 2],
+    "lbrace": "off",
+    "linebreak-style": ["error", "unix"],
+    "max-len": ["error", 79],
+    "no-constant": ["error"],
+    "no-empty-blocks": "off",
+    "quotes": ["error", "double"],
+    "uppercase": "off",
+    "visibility-first": "error",
+
+    "security/enforce-explicit-visibility": ["error"],
+    "security/no-block-members": ["warning"],
+    "security/no-inline-assembly": "off"
+  }
+}
