+++ docs/scripts/linenumber.js
@@ -21,4 +21,4 @@
       }
     }
   }
-})();
+}());
