+++ package.json
@@ -15,7 +15,8 @@
     "ethereumjs-wallet": "^0.6.3",
     "ganache-cli": "^6.2.3",
     "lowdb": "^1.0.0",
-    "truffle": "^4.1.14",
+    "truffle": "^5.0.0",
+    "truffle-hdwallet-provider": "^1.0.0-web3one.5",
     "truffle-wallet-provider": "0.0.5",
     "web3": "^1.0.0-beta.36",
     "web3-eth-abi": "^1.0.0-beta.36",
@@ -34,18 +35,26 @@
     "ethereumjs-util": "^6.0.0",
     "mocha": "^5.2.0",
     "nyc": "^13.1.0",
+    "rimraf": "^2.6.3",
     "sinon": "^7.2.2",
     "solidity-coverage": "^0.5.11"
   },
   "scripts": {
-    "test": "NODE_ENV=TEST mocha ./demo --trace-warnings --exit --colors --recursive --reporter spec",
-    "testCoverage": "NODE_ENV=TEST nyc mocha ./demo --trace-warnings --exit --colors --recursive --reporter spec",
-    "doorbellTest": "NODE_ENV=TEST mocha ./extractPublicKey --trace-warnings --exit --colors --recursive --reporter spec",
-    "demoRinkeby": "NODE_ENV=RINKEBY node ./demo/demoTransactions.js",
-    "demoRopsten": "NODE_ENV=ROPSTEN node ./demo/demoTransactions.js",
-    "demoKovan": "NODE_ENV=KOVAN node ./demo/demoTransactions.js",
-    "demoDevelopment": "NODE_ENV=DEVELOPMENT node ./demo/demoTransactions.js",
-    "demoMainnet": "NODE_ENV=MAINNET node ./demo/demoTransactions.js"
+    "clean": "./node_modules/.bin/rimraf -rf ./coverage.json ./build ./coverage ./coverageEnv ./export",
+    "lint:js": "./node_modules/.bin/eslint .",
+    "lint:js:fix": "./node_modules/.bin/eslint . --fix",
+    "lint:sol": "./node_modules/.bin/solium -d .",
+    "lint:sol:fix": "./node_modules/.bin/solium -d . --fix",
+    "lint": "npm run lint:js && npm run lint:sol",
+    "lint:fix": "npm run lint:js:fix && npm run lint:sol:fix",
+    "test": "NODE_ENV=TEST ./node_modules/.bin/mocha ./demo --trace-warnings --exit --colors --recursive --reporter spec",
+    "test:coverage": "NODE_ENV=TEST ./node_modules/.bin/nyc ./node_modules/.bin/mocha ./demo --trace-warnings --exit --colors --recursive --reporter spec",
+    "doorbell:test": "NODE_ENV=TEST ./node_modules/.bin/mocha ./extractPublicKey --trace-warnings --exit --colors --recursive --reporter spec",
+    "demo:rinkeby": "NODE_ENV=RINKEBY node ./demo/demoTransactions.js",
+    "demo:ropsten": "NODE_ENV=ROPSTEN node ./demo/demoTransactions.js",
+    "demo:kovan": "NODE_ENV=KOVAN node ./demo/demoTransactions.js",
+    "demo:development": "NODE_ENV=DEVELOPMENT node ./demo/demoTransactions.js",
+    "demo:mainnet": "NODE_ENV=MAINNET node ./demo/demoTransactions.js"
   },
   "repository": {
     "type": "git",
